pyproject.toml
[project]
name = "HousingModel"
version = "0.1.0"
description = "HousingModel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic",
    "pydantic-settings",
    "SQLAlchemy",
    "alembic",
    "geoalchemy2",
    "pyarrow",
    "duckdb",
    "pandera",
    "structlog",
    "prometheus-client",
    "requests-cache",
    "tenacity",
    "prefect",
    "geopandas",
    "shapely",
    "pyproj",
    "feedparser",
    "beautifulsoup4",
    "pandas",
    "hypothesis",
    # GUI dependencies
    "fastapi",
    "uvicorn[standard]",
    "gunicorn",
    "dash",
    "dash-bootstrap-components",
    "plotly",
    "python-multipart",  # for file uploads
    "jinja2",  # for templates
    "aiofiles",  # for async file operations
    # Mapping and visualization dependencies
    "dash-leaflet",  # Interactive mapping
    "geopandas",     # Geographic data handling (already present)
    "folium",        # Alternative mapping option
    "mapboxgl",      # Mapbox integration
    "docxtpl",       # Word memo templating
    "branca",        # Map styling utilities
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [
    "demo_complete_aker_integration.py",
    "test_optimize_cadence_simple.py",
    "test_state_packs_simple.py",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "."
ignore_missing_imports = true
